First step! Can actually pull data from .yaml files.
Getting here was a bit of a journey! The end result was that I
discovered that although each subfield (specifically the sub-fields of
`body`) doesn't have to be on each parent field, subfields DO need to
have matching types: had strings for most `content` subfields, but one
had an array--this was preventing GraphQL from being able to query them,
(although this was not communicated in any error messages). By changing
up how it works (specifically, I modified the schema for the `list`
type), everything is now queryable.